django_simple_crud_rest_api/api/serializers/course_serializer.py
from typing import Any
from rest_framework import serializers

from api.models.course import Course

class CourseSerializer(serializers.ModelSerializer):
    """
    Serializer for Course model.
    """
    class Meta:
        model = Course
        fields = '__all__'