django_simple_crud_rest_api/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.2
Django==5.2.2
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
idna==3.10
iniconfig==2.1.0
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.11.1
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
typing_extensions==4.14.0
urllib3==2.4.0
